move advanced panel sections into dedicated sections folder
