+++ src/components/panels/zones/CustomStopZoneSection.tsx
@@ -9,7 +9,7 @@ import {
   ToggleBtn,
   CardDivider,
   InfoIcon,
-} from "../advanced/shared";
+} from "../advanced/sections/shared";
 
 interface Props {
   settings: Settings;
